move multirangetraversal logic to helper
